requirements.txt
pymilvus==2.3.4
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
networkx==3.1
scikit-learn==1.3.0
python-dotenv==1.0.0
tqdm==4.66.1
matplotlib==3.7.2
seaborn==0.12.2
accelerate==0.24.1
huggingface-hub==0.17.3
